# learning-platform/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.UserResponse:
    properties:
      avatarUrl:
        type: string
      displayName:
        type: string
      email:
        type: string
      id:
        type: string
      role:
        type: string
    type: object
...
  response.ErrorMessage:
    properties:
      message:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorMessage'
      success:
        example: false
        type: boolean
    type: object
  response.SuccessWrapper:
    properties:
      data: {}
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for Learning Platform — users, auth, tasks, topics
  title: Learning Platform API
  version: "1.0"
paths:
...
  /user/all:
    get:
      description: Returns list of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessWrapper'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.UserResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - users
  /user/profile:
    get:
      description: Returns profile of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessWrapper'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - users
    put:
      consumes:
      - multipart/form-data
      description: Updates profile fields and optionally uploads avatar
      parameters:
      - description: Updated email
        in: formData
        name: email
        type: string
      - description: Updated display name
        in: formData
        name: displayName
        type: string
      - description: Avatar image
        in: formData
        name: avatar
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessWrapper'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - users
swagger: "2.0"